OfficeBench/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asttokens 
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
azure-identity==1.22.0
beartype==0.12.0
beautifulsoup4==4.12.3
blinker==1.9.0
browsergym==0.13.3
browsergym-assistantbench==0.13.3
browsergym-core==0.13.3
browsergym-experiments==0.13.3
browsergym-miniwob==0.13.3
browsergym-visualwebarena==0.13.3
browsergym-webarena==0.13.3
browsergym-workarena==0.4.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm 
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
debugpy 
decorator 
dill==0.3.8
distro==1.9.0
docker==7.1.0
docx2pdf==0.1.8
easyocr==1.7.2
editdistance==0.8.1
efficientnet==1.0.0
english-words==2.0.1
essential-generators==1.0
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup
executing 
Faker==33.1.0
Farama-Notifications==0.0.4
filelock==3.16.1
fire==0.5.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
gymnasium==1.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
icalendar==6.1.0
idna==3.10
imageio==2.36.1
imgaug==0.4.0
importlib_metadata
ipykernel
ipython
itsdangerous==2.2.0
jedi
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jupyter_client
jupyter_core
keras==3.7.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
libvisualwebarena==0.0.15
libwebarena==0.0.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.0
matplotlib-inline 
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msal==1.31.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
namex==0.0.8
nest_asyncio 
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.57.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging 
pandas==2.0.3
parso 
pdf2docx==0.5.8
pexpect 
pickleshare 
pillow==11.0.0
platformdirs 
playwright==1.49.1
plumbum==1.9.0
portalocker==3.0.0
prompt_toolkit
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil 
ptyprocess
pure_eval 
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyee==12.0.0
Pygments 
PyJWT==2.10.1
PyMuPDF==1.24.14
pyparsing==3.2.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil 
python-docx==1.1.0
pytz==2024.2
PyYAML==6.0.2
pyzmq 
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpyc==6.0.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacrebleu==2.4.3
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.3.2
scipy==1.14.1
shapely==2.0.6
six 
sniffio==1.3.1
soupsieve==2.6
stack_data
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
text-generation==0.7.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tornado
tqdm==4.67.1
traitlets 
transformers==4.47.1
triton==3.1.0
types-requests==2.32.0.20241016
types-tqdm==4.67.0.20241119
typing-inspect==0.9.0
typing_extensions 
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
wcwidth
weblinx==0.3.2
weblinx-browsergym==0.0.1.dev14
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp 